for3.4/requirements.txt
aiohttp==0.21.6
chardet==2.3.0
pyserial==3.0.1
